+++ tx-backend/src/main/resources/db/migration/V17__set_30_chars_for_import_state.sql
@@ -1,0 +1,2 @@
+ALTER TABLE assets_as_planned ALTER COLUMN import_state TYPE VARCHAR (30);
+ALTER TABLE assets_as_built ALTER COLUMN import_state TYPE VARCHAR(30);