+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/importdata/ImportControllerIT.java
@@ -431,7 +431,7 @@ class ImportControllerIT extends IntegrationTestSpecification {
         eventually(() -> {
             AssetBase asset = assetAsBuiltRepository.getAssetById("urn:uuid:254604ab-2153-45fb-8cad-54ef09f4080f");
             assertThat(asset.getPolicyId()).isEqualTo("default-policy");
-            assertThat(asset.getImportState()).isEqualTo(ImportState.PUBLISHED_TO_CX);
+            assertThat(asset.getImportState()).isEqualTo(ImportState.PUBLISHED_TO_CORE_SERVICES);
             dtrApiSupport.verifyDtrCreateShellCalledTimes(1);
             return true;
         });
@@ -477,7 +477,7 @@ class ImportControllerIT extends IntegrationTestSpecification {
             try {
                 AssetBase asset = assetAsBuiltRepository.getAssetById("urn:uuid:254604ab-2153-45fb-8cad-54ef09f4080f");
                 assertThat(asset.getPolicyId()).isEqualTo("default-policy");
-                assertThat(asset.getImportState()).isEqualTo(ImportState.PUBLISHED_TO_CX);
+                assertThat(asset.getImportState()).isEqualTo(ImportState.PUBLISHED_TO_CORE_SERVICES);
                 dtrApiSupport.verifyDtrCreateShellCalledTimes(1);
             } catch (AssertionFailedError exception) {
                 return false;