+++ CHANGELOG.md
@@ -11,12 +11,12 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 
 ### Added
 - #515 Service Unavailable Response on Notification failure
-- #536 Added import state PUBLISHED_TO_CX in frontend
+- #536 Added import state PUBLISHED_TO_CORE_SERVICES in frontend
 - #420 add /contracts api to fetch contract agreement information from EDC for assets
 - Added a step to the pull-request-backend.yml which checks if the pom.xml(root) properties have some versions ending with -SNAPSHOT
 - Added a PostConstruct method in PolicyStartUpConfig to allow Integration tests to run without errors in stack traces.
-- #536 added new ImportState to asset PUBLISHED_TO_CX indicating edc assets and dtr shells were created for given asset
-- #536 added cron job responsible to publish assets in PUBLISHED_TO_CX import state to edc and dtr
+- #536 added new ImportState to asset PUBLISHED_TO_CORE_SERVICES indicating edc assets and dtr shells were created for given asset
+- #536 added cron job responsible to publish assets in PUBLISHED_TO_CORE_SERVICES import state to edc and dtr
 
 ### Changed
 - Updated RELEASE.md to the latest release guide (added more steps)